* enc/utf_8.c (code_to_mbc): suppressed a warning.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@24607 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ enc/utf_8.c
@@ -363,7 +363,7 @@ code_to_mbc(OnigCodePoint code, UChar *buf, OnigEncoding enc ARG_UNUSED)
     }
 
     *p++ = UTF8_TRAIL0(code);
-    return p - buf;
+    return (int)(p - buf);
   }
 }
